T F -> 1 0 (#32)
* true false => 1 0 change

etc dto, repository fix

* create-drop -> update

+++ src/main/java/com/ict/traveljoy/plan/progress/repository/PlanProgress1.java
@@ -52,7 +52,7 @@ public class PlanProgress1 {
     private Integer travelers;
 
     @Column(name = "TRAVEL_COST")
-    private BigDecimal travelCost;
+    private Long travelCost;
 
     @ManyToOne
     @JoinColumn(name = "PLAN_ID")
